Separating Carousels code. Fixing reverse sliding in MENA.
